fix for 32 bit machines

+++ model/Package_bitmessaged/Package_data/classes/class_node_info/operations/operation_TGVSGRKO/code.cpp
@@ -5,7 +5,7 @@ obj["ip"] = inet_ntoa(myAddr.sin_addr);
 obj["port"] = (uint64_t)myAddr.sin_port;
 
 if (currentlyInUse > 0)
-    obj["last_connected"] = std::time(0);
+    obj["last_connected"] = (uint64_t)std::time(0);
 else
     obj["last_connected"] = lastConnectedTime;
